+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 0.4.0 [unreleased]
 
 - Reduced the number of dependencies from 42 to 9
+- Added support for Bioconductor remotes
 
 Breaking changes